+++ docs/conf.py
@@ -27,7 +27,7 @@ sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../'
 
 # Add any Sphinx extension module names here, as strings. They can be extensions
 # coming with Sphinx (named 'sphinx.ext.*') or your custom ones.
-extensions = ['cornice.ext.sphinxext', 'sphinx.ext.autodoc', 'sphinx.ext.extlinks',
+extensions = ['cornice_sphinx', 'sphinx.ext.autodoc', 'sphinx.ext.extlinks',
               'sphinx.ext.todo', 'sphinx.ext.coverage', 'sphinx.ext.viewcode']
 
 # Add any paths that contain templates here, relative to this directory.